delete `stories` dir, add example Button component
