+++ .storybook/main.ts
@@ -2,8 +2,7 @@ import type { StorybookConfig } from "@storybook/react-webpack5";
 
 const config: StorybookConfig = {
   stories: [
-    "../stories/**/*.mdx",
-    "../stories/**/*.stories.@(js|jsx|mjs|ts|tsx)",
+    "../src/**/*.stories.tsx"
   ],
   addons: [
     "@storybook/addon-webpack5-compiler-swc",
