+++ src/components/Button/Button.tsx
@@ -27,7 +27,7 @@ export interface ButtonProps {
 /**
  * Primary UI component for user interaction
  */
-export const Button = ({
+const Button = ({
   primary = false,
   size = 'medium',
   backgroundColor,
@@ -46,3 +46,5 @@ export const Button = ({
     </button>
   );
 };
+
+export default Button;
